# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notetools API",
        "version": "2.0.1",
        "description": "API for de gerenciamento de pedidos"
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "nt.authtoken"
            }
        },
...
    },
    "paths": {
...
        "/insights/{userId}": {
            "get": {
                "tags": ["Insights"],
                "summary": "Obter insights do usuário",
                "description": "Obtém insights sobre o usuário, como pedidos pendentes, notas e status de acesso",
                "security": [
                    {
                        "cookieAuth": ["nt.authtoken"]
                    }
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "params",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dateIn": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data inicial para filtrar os insights"
                                    },
                                    "dateOut": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data final para filtrar os insights"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Insights obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "revenue": {
                                            "type": "number",
                                            "description": "Receita total do usuário"
                                        },
                                        "sales": {
                                            "type": "number",
                                            "description": "Quantidade total de pedidos do usuário"
                                        },
                                        "completionRate": {
                                            "type": "number",
                                            "description": "Taxa de conclusão de pedidos do usuário"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token não localizado"
                    },
                    "404": {
                        "description": "Usuário não localizado"
                    }
                }
            }
        },
        "/insights/status/{userId}": {
            "get": {
                "tags": ["Insights"],
                "summary": "Obter status de pedidos do usuário",
                "description": "Obtém o status dos pedidos de um usuário específico",
                "security": [
                    {
                        "cookieAuth": ["nt.authtoken"]
                    }
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "params",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dateIn": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data inicial para filtrar os insights"
                                    },
                                    "dateOut": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data final para filtrar os insights"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status dos pedidos obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "pending": {
                                            "type": "integer",
                                            "description": "Número de pedidos pendentes"
                                        },
                                        "connected": {
                                            "type": "integer",
                                            "description": "Número de pedidos conectados"
                                        },
                                        "cancelled": {
                                            "type": "integer",
                                            "description": "Número de pedidos cancelados"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token não localizado"
                    },
                    "404": {
                        "description": "Usuário não localizado"
                    }
                }
            }
        },
        "/insights/per-day/{userId}": {
            "get": {
                "tags": ["Insights"],
                "summary": "Obter pedidos por dia do usuário",
                "description": "Obtém a quantidade de pedidos feitos por dia de um usuário específico",
                "security": [
                    {
                        "cookieAuth": ["nt.authtoken"]
                    }
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "params",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dateIn": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data inicial para filtrar os pedidos"
                                    },
                                    "dateOut": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "Data final para filtrar os pedidos"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedidos por dia obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "sales": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "day": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token não localizado"
                    },
                    "404": {
                        "description": "Usuário não localizado"
                    }
                }
            }
        }
    }
}